requirements.txt
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-import-export==1.2.0
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.0
pytz==2019.2
PyYAML==5.1.2
sqlparse==0.3.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0
boto3==1.5.32
botocore==1.8.46
psycopg2==2.7.3.2